moved economy command validation helper, updated imports
